新增：游戏启动可视

+++ windows/sql/data_structs.h
@@ -68,7 +68,7 @@ struct SubjectRelationData {
 struct GameData {
     int subjectId = 0;
     QString launchPath;
-    double playDuration = 0.0;
+    int playDuration = 0;
 };
 
 Q_DECLARE_METATYPE(EpisodeData)
